Change field type to match program model
To help simplify model updates to the database, we sacrifice some
memory to reduce looping cycles. However since other components
have not been updated, we had to add a helper/adapter method.

An additional database operation had be written as part of the
adapter. This is also the first part of the database refactoring